Clean up dependencies
